Refactor verifier so it can be tested completely hermetically
