+++ poc_mobile_verifier/src/cli/mod.rs
@@ -1,2 +1,6 @@
 pub mod generate;
 pub mod server;
+
+pub const CONNECT_TIMEOUT: std::time::Duration = std::time::Duration::from_secs(5);
+pub const RPC_TIMEOUT: std::time::Duration = std::time::Duration::from_secs(5);
+pub const DEFAULT_URI: &str = "http://127.0.0.1:8080";
